Fix the world Add titles to error messages loooooooooooooong currency

+++ DiscordHackWeek2019/DiscordHackWeek2019/Commands/BotCommandContext.cs
@@ -105,7 +105,7 @@ namespace DiscordHackWeek2019.Commands
         public InventoryWrapper GetInventory(ulong user) => new InventoryWrapper(this, user);
         public InventoryWrapper GetInventory(User user) => new InventoryWrapper(this, user);
 
-        public string Money(int amount) => $"e̩̍{amount}";
+        public string Money(long amount) => $"e̩̍{amount}";
 
         public void ClearCachedValues()
         {
