+++ DiscordHackWeek2019/DiscordHackWeek2019/Config/Options.cs
@@ -7,6 +7,6 @@ namespace DiscordHackWeek2019.Config
     [ConfigFile("Config/options.json")]
     public class Options : Config
     {
-        public int StartingCurrency { get; set; }
+        public long StartingCurrency { get; set; }
     }
 }
